ACTIVERECORD_JDBC-156: tentative patch -- fix type error "undefined method 'first' for false:FalseClass" when making include? query on has_many :through association which is not yet loaded
This is a tentative patch -- it fixes code which seemingly cannot be correct by removing the possibility of a type error, but I don't know enough about ARel to say that its semantics are correct.  In addition, when it is applied and you make the same query, a second problem arises: the generated query now fails when submitted to the SQL server because it asks for a column twice.