Naive attempt to support ARel 2.2
The most important change works for all adapters: ARel now passes a
TreeManager instead of the parsed SQL to adapters; This was handled by
calling #to_sql on the manager (if the object actually responds to that)
deep inside #substitute_parameters, which should cover all types of
statements.

Also, the global Arel::Visitors::VISITORS is deprecated, and instead
each adapter should respond to :visitor_for returning an appropriate
visitor. The MySQL adapter was changed to do that, but other adapters
are still using the deprecated API.