disable extension auto-discovery for installed gems (unless specified)
probably invented back the pre-bundler days, there's definitely issues with scanning all (installed) gems e.g. when another version of the AR-JDBC gem is installed it will call it's dicover.rb file.
it should no longer be needed for extensions since they are likely to be specified in a Gemfile and auto-loaded - a correct extension will thus load as expected even without the (disabled) discovery mechanism ...
for compatibility we still support discovery with a system property