[postgres]: Improve handling of DB type "money"
When using BigDecimal for money values, the standard handling with
setPGobjectParameter() does not work since it does a toString() on the
value. For a BigDecimal("123.45") this will produce "0.12345e3" which
PG doesn't like.

This registers a handler for PG type "money" which uses
statement.setBigDecimal() when a RubyBigDecimal is provided and falls
back to the standard handling setPGobjectParameter() for everything
else.

This fixes test errors:
* ajrdbc: test/db/postgresql/types_test.rb:608 in 'test_update_money'
* rails: test/cases/adapters/postgresql/money_test.rb:68:in `test_create_and_update_money'