[mysql] make fixtures work and Rails 5.2 test run again (#909)
Rails 5.2 introduced a insert_fixtures_set() with special optimization
for MySQL, namely using multiple statements in one call. This requires
the connection params to be set up and a method with_multi_statements
that we're not (yet) implementing. The C version changes flags of the
connection for that.

The easy fix is to just use the plain loop from the SQLite3 adapter.
This is slower than it could be, but at least the Rails tests run again
with this. And it looks pretty nice already:

5459 runs, 15416 assertions, 15 failures, 6 errors, 5 skips

(that's with jruby-9.2 so two of the failures are about FrozenError..)