[52, postgres, RFH] Fix Rails date tests with year 0, BC, infinity
Infinity handling is just the same as with Timestamp.

Dates with year 0 and BC is a bit more problematic since Java,
Ruby and PG all have a different chronology.

It's down to ISO/Gregorian/Julian chronologies and a mix thereof...
Writing a date with year 0/BC does work with PG, but reading it returns
something else, year 0 becomes -1. So when reading, it's necessary to
parse the "BC" in the date string _and_ use a GJChronology.

Fixes (with JRuby 9.1.17):
* rails: test/cases/adapters/postgresql/date_test.rb
=> 5 errors or so less in total...

This will produce errors in JRuby 9.2.0 but will work in 9.2.1