Travis: use postgresql-9.4
By default Travis uses PG 9.2. But it seems the client is actually newer
than that. One of the test errors is the server complaining about
"lock_timeout" which is a 9.3+ thing. Using a newer server version
fixes the problem. 9.2 is EOL anyway and 9.3 will be in a few weeks,
so go for 9.4