Perf: Log raw bind values instead of type casted ones
The MRI AR adapters log the type casted values. Logging the raw values
is faster and since ARJDBC does casting internally using Java code where
necessary, it doesn't make much sense to log casted values when in fact
it isn't what is passed to JDBC.

The downside is that it might be unexpected to see e.g. a symbol instead
of a string in the log...