[postgres] support timestamp arrays
pgjdbc uses toString() on each value of a timestamp array. W/o any special
treatment, this means on RubyTime objects. Since the string format is not
exactly what PG supports, precision is off, BC dates are not supported.
And time zones are not working. Using java.sql.Timestamp or even
PGTimestamp does no good either. Implement a RubyTime -> String helper
and use that. This is based on what pgjdbc does internally, but adopted
for JodaTime.

Fixes:
* rails/test/cases/adapters/postgresql/array_test.rb