[test] exclude multi-db tests for now
These tests all try various multi-DB things, using an SQLite3 database. When
running the Rails tests from AR-JDBC, the SQLite3 JARs are not on the search
path so this fails.

At a later point, the search path can be changed to include SQLite3 JARs,
but for now this ensures we're actually testing PG/MySQL

This bring the Rails test from 600+ errors down to a reasonable
  PostgreSQL: 34 failures, 1 errors
  MySQL: 31 failures, 0 errors