[postgres] exclude test that fails for different return value of execute()
execute() is supposed to return the raw(-ish) value from the database call.
Basically a statement can have three different kinds of output:
* a set of rows (ResultSet in JDBC)
* an update count for insert/update/delete statements
* nothing at all

In JDBC, it's only possible to distinguish between a ResultSet and an
update count, so a "SET ..." returns 0, while the pg gem return an empty
array. Exclude the test since there isn't an easy fix for this.