Add some PG::Result compatibility methods.  Add a typemap registration system similar in spirit to Result.typemap= but instead put it on connection (our resultset.close() logic makes emulating putting on result very invasive).
This work will allow people who are breaking out of AR to just use PG the
ability to do that.  Discourse does this with their gem mini_sql which I have
a fork I am hoping to merge soon.

There should be no performance impact from this as the typemap is not used by
AR and will only pay a single null check.  I also outlined the logic for
applying the typemap so inlining should not be a concern.

I do need to add some tests but I made a marker interface PG to point out that
these methods are intended as emulation methods.  I suspect as more data is run
through mini_sql it may lead to a nice set of specific tests.