[perf] introduce option to set JDBC fetch size
For some workloads and databases it can be beneficial to set a JDBC fetch size
to increase performance and decrease memory usage. This can help a lot (5x and
more in terms of speed) with result sets withs thousands of rows or more.

The feature is disabled by default and can be enabled by adding the config
option "jdbc_fetch_size", set to the desired number of rows (a value between
100 and 1000 is usually best, but depends on the workload). Since this
is just a hint to the underlying JDBC driver, the option might only work
in combination with other properties or not at all.

While at it, always close ResultSet right after use. This shouldn't matter
but there is no reason to special case anything.