[core] throw an error when unable to get a connection
When a wrong URL is provided, it can happen that newConnection() returns
null. If this happens during configure_connection, the whole process is
called over and over again and results in a stack overflow.
Instead, just throw an error.

Noticed because a new test in Rails master tries to connect with
"mysql2:///wait_timeout=60", an URL unknown to JDBC.