[core] use RubyDateTime directly to create a string
With Rails 6.1, there's no conversion to a Time anymore for a DateTime
if it's already UTC. So what happens is that we call #to_time where
we dind't before. Since Rails overrides that for DateTime to basically
do #getlocal, the result is off.

Using RubyDateTime directly avoids the problem and is faster anyway.